S-E-87 [PORT] efarmer_sale_workflow
Port to v15.

ir.actions.server:activity_record is not used on the prod,
so there is no reason to support it.

Assets should be declared in manifest now.

Clean layout doesn't exist anymore.
Light layout is expected to be used.

account.move:type => account.move:move_type

It's better to throw UserError to display
the error message properly.